+++ src/V1/Assurances/ComposableAssurance.php
@@ -43,16 +43,18 @@
 
 namespace GanbaroDigital\Defensive\V1\Assurances;
 
-use GanbaroDigital\Defensive\V1\Exceptions\BadAssurance;
-use GanbaroDigital\Defensive\V1\Exceptions\BadAssuranceArgs;
 use GanbaroDigital\Defensive\V1\Exceptions\DefensiveExceptions;
 use GanbaroDigital\Defensive\V1\Interfaces\Assurance;
+use GanbaroDigital\Defensive\V1\Interfaces\ListAssurance;
 use GanbaroDigital\DIContainers\V1\Interfaces\FactoryList;
 
-class ComposableAssurance implements Assurance
+class ComposableAssurance implements Assurance, ListAssurance
 {
     // saves us having to declare ::__invoke() ourselves
     use InvokeableAssurance;
+
+    // saves us having to declare ::toList() ourselves
+    use ListableAssurance;
 
     /**
      * the partial assurance that we are converting into a